+++ app/assets/scripts/selfie/directives.js
@@ -3,7 +3,24 @@ function( angular ) {
     'use strict';
 
     var $ = angular.element,
-        extend = angular.extend;
+        extend = angular.extend,
+        copy = angular.copy,
+        forEach = angular.forEach,
+        isObject = angular.isObject;
+
+    function deepExtend(target, extension) {
+        forEach(extension, function(extensionValue, prop) {
+            var targetValue = target[prop];
+
+            if (isObject(extensionValue) && isObject(targetValue)) {
+                deepExtend(targetValue, extensionValue);
+            } else {
+                target[prop] = copy(extensionValue);
+            }
+        });
+
+        return target;
+    }
 
     return angular.module('c6.app.selfie.directives', [])
         .directive('c6FillCheck', ['$timeout',
@@ -112,6 +129,108 @@ function( angular ) {
                     });
                 }
             };
+        }])
+
+        .directive('selfiePreview', [function() {
+            return {
+                restrict: 'E',
+                scope: {
+                    card: '=',
+                    active: '=?',
+                    device: '=?',
+                    profile: '=?',
+                    experience: '=?',
+                    standalone: '=?'
+                },
+                templateUrl: 'views/selfie/directives/preview.html',
+                controller: 'SelfiePreviewController',
+                controllerAs: 'SelfiePreviewCtrl'
+            };
+        }])
+
+        .controller('SelfiePreviewController', ['$scope','MiniReelService','c6BrowserInfo',
+        function                               ( $scope , MiniReelService , c6BrowserInfo ) {
+            var SelfiePreviewCtrl = this,
+                experience;
+
+            // we call this when the card is changed or when the
+            // base experience is overridden
+            function loadPreview(card) {
+                if (!card) { return; }
+
+                MiniReelService.convertCardForPlayer(card)
+                    .then(function(cardForPlayer) {
+                        var newExperience = copy(experience);
+
+                        cardForPlayer.data.autoplay = false;
+                        cardForPlayer.data.skip = true;
+                        cardForPlayer.data.controls = true;
+
+                        newExperience.data.deck = [cardForPlayer];
+
+                        SelfiePreviewCtrl.card = cardForPlayer;
+                        SelfiePreviewCtrl.experience = newExperience;
+                    });
+            }
+
+            // on initiation we create a base experience for previewing
+            // this can be overridden on the scope by the consumer
+            MiniReelService.create()
+                .then(function(minireel) {
+                    experience = deepExtend(minireel, {
+                        id: 'e-123',
+                        data: {
+                            mode: 'light',
+                            autoplay: false,
+                            autoadvance: false,
+                            adConfig: {
+                                video: {
+                                    firstPlacement: -1,
+                                    frequency: 0
+                                },
+                                display: {}
+                            }
+                        }
+                    });
+                });
+
+            // we allow the consumer to set this on initiation
+            // but default to the browser info
+            // maybe we want to allow more dynamic updating?
+            this.profile = $scope.profile || copy(c6BrowserInfo.profile);
+
+            // we allow the consumer to set these, we default to true
+            this.active = typeof $scope.active === 'undefined' || $scope.active;
+            this.standalone = typeof $scope.standalone === 'undefined' || $scope.standalone;
+
+            // we allow the consumer to this, we default to desktop
+            // we're watching this value so the consumer can change
+            // outside the directive
+            $scope.device = $scope.device || 'desktop';
+
+            // watch for changes to the campaign's card
+            $scope.$watch('card', loadPreview);
+
+            // watch for changes to the experience in case
+            // the consumer wants to mess with mode or whatnot
+            $scope.$watch('experience', function(exp) {
+                if (!exp) { return; }
+                experience = copy(exp);
+                loadPreview($scope.card);
+            });
+
+            // watch for changes in case the consumer wants to
+            // override the device outside of the directive
+            $scope.$watch('device', function(device) {
+                var profile = SelfiePreviewCtrl.profile;
+
+                if (device === profile.device) { return; }
+
+                SelfiePreviewCtrl.profile = extend(copy(profile), {
+                    device: device,
+                    flash: device !== 'phone'
+                });
+            });
         }])
 
         .filter('videoService', [function() {